fix(go-cli): decode attachments in `durian attachment --save`
The CLI --save path called FetchBodySection directly, writing raw
transfer-encoded bytes (base64 / quoted-printable) to disk while
reporting the decoded size from the attachments table. For
multipart-nested attachments it also misused the DB part_id as a
top-level IMAP section number, so the on-disk file was a chunk of the
surrounding multipart body instead of the attachment.

The HTTP attachment route already handled this correctly via the
watcher fetch path. Extract that logic into reusable imap.Client
helpers and share between watcher and CLI:

- new cli/internal/imap/attachment.go: FetchDecodedAttachment,
  FetchAndDecodeBodySection, FindAttachmentSection, IsAttachmentLike,
  plus DFS walkers. walkForIndex now returns []int{1} for a top-level
  single-part body (RFC 3501 §6.4.5) so DMARC-style application/zip
  messages can be fetched.
- cli/cmd/durian/attachment.go: call FetchDecodedAttachment, report
  the actual written byte count instead of the DB-recorded size.
- cli/internal/handler/watcher.go: call the shared helper, drop the
  duplicated decode switch and BODYSTRUCTURE walkers.
- Tests for IsAttachmentLike, walk*, FindAttachmentSection move from
  cli/internal/handler/ to cli/internal/imap/.

Fixes the second half of #227.
